requirements.txt
botcity-slack-plugin